requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
cgi-tools==0.0.4
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.2.0
googletrans==3.1.0a0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.6.9
idna==2.10
inflection==0.5.1
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
legacy-cgi==2.6.3
loguru==0.7.3
MarkupSafe==3.0.2
mccabe==0.7.0
mock==5.2.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
simplejson==3.20.1
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
whitenoise==6.9.0
